+++ ChangeLog.md
@@ -13,7 +13,7 @@ ClickHaskell documentation got it's own domain name: https://clickhaskell.dev/
 - Dropped vector dependency
 - Introduced memory consumption test (64M limit) on parallel reading and writing of 1 million rows
 - Added new reading wrapper for generateRandom function
-- Depricating `Ch*` prefixes on types `ChUInt*` (to Word* type synonyms), `ChInt*`(to Int* standart types), `ChDate*` (to Date and DateTime)
+- Depricating `Ch*` prefixes on types `ChUInt*` (to Word* type synonyms), `ChInt*`(to Int* standart types), `ChDate*` (to Date and DateTime type), `ChArray` (to Array type), `ChUUID` (to UUID type)
 - openNativeConnection now passes $HOME and $USERNAME variables to query info
 
 ## Breaking changes: